+++ src/features/psionics/psionics.js
@@ -1,6 +1,7 @@
 import { addPsionicSubtitles } from './ui-spellbook';
 import { reorganizePsionicDisciplines } from './ui-features';
 import { registerPsychicFocusManagement } from './psychic-focus';
+import { registerArgonIntegration } from './ui-argon';
 
 function setupPsionicOptions() {
     CONFIG.DND5E.featureTypes.discipline = {
@@ -58,4 +59,5 @@ export function registerPsionics() {
     Hooks.on('renderActorSheet5eCharacter2', reorganizePsionicDisciplines);
     Hooks.on('renderActorSheet5eNPC2', reorganizePsionicDisciplines);
     registerPsychicFocusManagement();
+    registerArgonIntegration();
 }